pyproject.toml
[project]
name = "pishro-lib"
version = "0.1.0"
description = "Pishro is a package manager and deployment tool for docker swarm"
readme = "README.md"
authors = [
    { name = "Siavash Outadi", email = "siavash.outadi@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "gitpython>=3.1.44",
    "jinja2>=3.1.6",
    "pydantic>=2.11.3",
    "python-on-whales>=0.76.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "types-docker>=7.1.0.20241229",
]
